% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_imgt_available}
\alias{is_imgt_available}
\title{Check if IMGT Website is Available}
\usage{
is_imgt_available()
}
\value{
A logical value: \code{TRUE} if the IMGT website is accessible,
  \code{FALSE} otherwise.
}
\description{
Sends a lightweight HEAD request to the main IMGT page to check
if the service is online and accessible. This function is used to
conditionally run examples and tests that require an internet connection.
}
\examples{
is_imgt_available()
}
\seealso{
\code{\link{is_ogrdb_available}} for checking OGRDB availability

\code{\link{getIMGT}} which uses this function
}
