% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_ogrdb_available}
\alias{is_ogrdb_available}
\title{Check if OGRDB Website is Available}
\usage{
is_ogrdb_available()
}
\value{
A logical value: \code{TRUE} if the OGRDB website is accessible,
  \code{FALSE} otherwise.
}
\description{
Sends a lightweight HEAD request to the OGRDB API to check if
the service is online and accessible. This function is used to conditionally
run examples and tests that require an internet connection.
}
\examples{
is_ogrdb_available()
}
\seealso{
\code{\link{is_imgt_available}} for checking IMGT availability

\code{\link{getOGRDB}} which uses this function
}
