% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{listIMGT}
\alias{listIMGT}
\title{List Datasets in the Local Cache}
\usage{
listIMGT()
}
\value{
A character vector of file paths for the cached datasets.
}
\description{
Scans the cache directory and returns a list of available datasets.
}
