% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_ogrdb_available}
\alias{is_ogrdb_available}
\title{Check if OGRDB Website is Available}
\usage{
is_ogrdb_available()
}
\value{
A logical value: `TRUE` if the IMGT website is accessible, `FALSE` otherwise.
}
\description{
This function sends a lightweight HEAD request to the main OGRDB page to
check if the service is online and accessible.
}
\examples{
  is_imgt_available()
}
