% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainOGRDB.R
\name{listOGRDB}
\alias{listOGRDB}
\title{List OGRDB Datasets in Local Cache}
\usage{
listOGRDB()
}
\description{
Scans the cache directory and returns a list of available datasets.
}
\examples{
if (is_ogrdb_available()) {
  # List cached OGRDB files
  cached_files <- listOGRDB()
  head(cached_files)
}
}
